requirements.txt
flask==1.1.2
gunicorn==20.0.4
python_dotenv==0.17.1
requests==2.25.1